requirements.txt
Flask==2.3.3
requests==2.31.0
Faker==19.6.2
openai==1.3.7
python-dotenv==1.0.0 